requirementFile.txt
fastapi
uvicorn
easyocr
openai
python-dotenv